feat: separate manager
